Refactored it sucessfully
