+++ src/client/cli.rs
@@ -1,0 +1,56 @@
+#![allow(unused_parens)]
+
+use clap::{
+    Parser,
+    Subcommand
+};
+use tokio::net::TcpStream;
+
+
+// Cli Parser
+#[derive(Parser)]
+#[command(version, about)]
+pub struct Cli {
+    #[command(subcommand)]
+    pub command: Option<Commands>,
+
+    #[arg(short, long, default_value_t = ("127.0.0.1:8080".to_string()) )]
+    pub addr: String
+}
+
+#[derive(Subcommand)]
+pub enum Commands {
+    /// Set a key-value pair
+    Set {
+        key: String,
+        val: String,
+        ttl: Option<u64>
+    },
+
+    /// Get the value for a key
+    Get { key: String },
+
+    /// Remove a key-value pair
+    Rm { key: String },
+
+    /// Increase the frequency of a key
+    IncrementFrequency { key: String },
+
+    /// Get the metadata for a key
+    GetMetadata { key: String },
+
+    /// Ping a IP address
+    Ping,
+
+    /// Get the size of the database
+    Size,
+
+    /// Flush the database
+    Flush
+}
+
+#[derive(Debug)]
+pub struct Client {
+    pub stream: TcpStream,
+    pub buf: Vec<u8>
+}
